requirements.txt
asgiref==3.8.1
dj-rest-auth==6.0.0
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
mysqlclient==2.2.4
python-slugify==8.0.4
sqlparse==0.5.0
text-unidecode==1.3
typing_extensions==4.11.0
